Alternate Page, Extreamly Minimal
Just a suggestion, Can be viewed in the folder 'Etchnew' Inside the cloned folder. I think we should go with a minimal template. And also there's a header aka Navigation pane missing in this template, otherwise its a simple single paged template to work with. One more thing, A new logo subject to corrections and suggestions is added, I've  included the svg file also. so if you want to edit it it'll be easy.