Added A neon background, Content update 0.3
Neon background is subject to the opinions. I don't think that the existing background is that much good. This image that I've just replaced the pevious one with is not the one perfect image because image quality is not that good. If we use an abstract svg, That'll be nice I think